shogi/pyproject.toml
[project]
name = "shogi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "python-shogi>=1.1.1",
    "streamlit>=1.56.0",
    "streamlit-image-coordinates>=0.4.0",
]
